requirements.txt
python-telegram-bot==12.8
readsettings==3.4.5
pymongo==3.6.0
dnspython==1.15.0
requests==2.24.0
pytz